Create rough console version of application
This is by no means an actually console-functional version of the app,
but I wanted to test out if the means of creating, accessing, and adding
projects/tasks worked with the current implementation. From my very
roughly sketched out console version, it seems like it mostly does. I
feel as though I've tested a good chunk of the core app logic, and it's
working, so instead of spending my time creating a really robust and
nuanced series of console prompts to go through all the possible ways
of interacting with the app, I should focus on creating the UI for the
functional logic that is far more intuitive and less reliant on
elaborate strings of reading input from the console.