Move dialog button event listener
I want to use the build dialog function to build a dialog for editing
already existing todos. I moved the adding of the event listener to
create a new todo to outside of the building of the dialog, allowing for
the ability to build the dialog for editing without the todo creating
event listener attached to it.